Refactor to handle multiple memcached instances
Without this change, only a single instance of memcached can be instantiated
and managed by Puppet on a given server.  This is a problem because it's
often important to use separate instances in order to safely use memcache to
store information from multiple sources (e.g. different applications or
different parts of the same application).

Due to the significance of this refactor, and the fact that I'm only
testing/using this on Ubuntu, I've stripped out support for other operating
systems (though it shouldn't be overly difficult to put them back in if
required - other than the memcached::instance defined type being added, the
API hasn't changed too dramatically).