Add a log directory with appropriate ownership
Without this change, memcached is not able to write to its logfiles if
started as a non-root user.  This change adds a log directory with
appropriate ownership, and makes the user a global setting rather than per
instance so that user can own the log directory and all instances can write
to it.